StyleCop: SA1649
