+++ StonehearthEditor/FileData.cs
@@ -6,13 +6,6 @@ using System.Windows.Forms;
 
 namespace StonehearthEditor
 {
-    public interface IModuleFileData
-    {
-        void SetModuleFile(ModuleFile moduleFile);
-
-        ModuleFile GetModuleFile();
-    }
-
     public abstract class FileData : IDisposable
     {
         protected TreeNode mTreeNode;
