+++ StonehearthEditor/IModuleFileData.cs
@@ -1,0 +1,9 @@
+﻿namespace StonehearthEditor
+{
+    public interface IModuleFileData
+    {
+        void SetModuleFile(ModuleFile moduleFile);
+
+        ModuleFile GetModuleFile();
+    }
+}
