+++ StonehearthEditor/IReloadable.cs
@@ -1,0 +1,7 @@
+﻿namespace StonehearthEditor
+{
+    public interface IReloadable
+    {
+        void Reload();
+    }
+}
