+++ StonehearthEditor/ManifestView.cs
@@ -2,21 +2,13 @@
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Drawing;
-using System.Data;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.IO;
 using Newtonsoft.Json.Linq;
 
 namespace StonehearthEditor
 {
-    public interface IReloadable
-    {
-        void Reload();
-    }
-
     public partial class ManifestView : UserControl, IReloadable
     {
         private FileData mSelectedFileData = null;
